Moves files between packages
