+++ src/test/java/com/codurance/allaboard/core/unit/WebAcceptanceTestTemplateShould.java
@@ -1,6 +1,6 @@
-package com.codurance.allaboard.web.utils;
+package com.codurance.allaboard.core.unit;
 
-import com.codurance.allaboard.core.unit.WebAcceptanceTestTemplate;
+import com.codurance.allaboard.web.utils.WebAcceptanceTestTemplate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.util.ReflectionTestUtils;
